loire_signer/requirements.txt
jsonschema
cryptography~=42.0.2
jwcrypto~=1.5.3
PyLD~=2.0.3
PyJWT~=2.8.0
uvicorn~=0.27.1
fastapi~=0.109.2
pydantic~=2.6.1
pandas~=2.1.1

